Create investment sub-app and render homepage  - Refactor to split export routes and investment routes.  - Move middleware out of lib  - Gerneral tidy up Remove 'm' unit from large number filter so it works better for volume
