+++ package.json
@@ -29,12 +29,12 @@
     "start:server": "node --use-strict server.js",
     "start:docker": "docker run -d -p 8080:8080 export-wins-ui-mi:${npm_package_version}",
     "start:docker-mac": "docker run -e \"COOKIE_SECRET=${COOKIE_SECRET}\" -e \"MI_SECRET=${MI_SECRET}\" -e \"MI_HOST=10.200.10.1\" -d -p 8080:8080 export-wins-ui-mi:${npm_package_version}",
-    "app": "node --use-strict src/app/app.js",
+    "app": "node --use-strict src/app/cluster.js",
     "watch-stub": "export STUB_MI=true; export FAKE_MI=false; npm run watch",
     "watch-fake-stub": "export STUB_MI=true; export FAKE_MI=true; npm run watch",
     "watch-mock-fake-stub": "export MOCK_MI=true; export STUB_MI=true; export FAKE_MI=true; npm run watch",
     "watch": "run-p watch:*",
-    "watch:node": "export LOG_LEVEL=debug; nodemon src/app/app.js --ignore 'src/public/**/*' --ignore 'src/test/**/*' --ignore 'src/data/**/*'",
+    "watch:node": "export LOG_LEVEL=debug; nodemon src/app/cluster.js --ignore 'src/public/**/*' --ignore 'src/test/**/*' --ignore 'src/data/**/*'",
     "build": "grunt dist",
     "build:docker": "docker build -t export-wins-ui-mi:${npm_package_version} .",
     "build:dist": "npm run sass:prod && grunt dist && npm run dist-install",
