+++ src/app/middleware/ping.js
