+++ src/app/sub-apps/investment/controllers/controller.index.js
@@ -1,0 +1,4 @@
+module.exports = function( req, res, next ){
+
+	res.render( 'investment/views/index' );
+};
