+++ src/app/sub-apps/investment/routes.js
@@ -1,0 +1,8 @@
+const indexController = require( './controllers/controller.index' );
+
+module.exports = function( router, isDev ){
+
+	router.get( '/', indexController );
+
+	return router;
+};
