+++ src/app/views/layout.html
@@ -21,9 +21,18 @@
 		</div>
 
 		<nav class="top-nav">
-			<a href="{{ urls.index() }}">MI Export wins</a>
+			<ul class="top-nav_list">
+				<li class="top-nav_list_item">
+					<a class="top-nav_link" href="{{ urls.index() }}">MI Export wins</a>
+				</li>
+				<li class="top-nav_list_item">
+					{% if showInvestment -%}
+					<a class="top-nav_link" href="{{ urls.investment.index() }}">MI Investment wins</a>
+					{% endif -%}
+				</li>
+			</ul>
 			{% if user.email -%}
-			<span class="username">User{% if user.internal %} (internal){% endif %}: {{ user.email }}</span>
+			<span class="top-nav_username">User{% if user.internal %} (internal){% endif %}: {{ user.email }}</span>
 			{%- endif %}
 		</nav>
